création AddressManager
