+++ src/Model/AddressManager.php
@@ -1,0 +1,33 @@
+<?php
+
+
+namespace App\Model;
+
+/**
+ *
+ */
+class AddressManager extends AbstractManager
+{
+    /**
+     *
+     */
+    const TABLE = 'address';
+
+    /**
+     *  Initializes this class.
+     */
+    public function __construct()
+    {
+        parent ::__construct(self::TABLE);
+    }
+    public function insertAddress(array $data): void
+    {
+        $statement = $this->pdo->prepare('INSERT INTO ' . self::TABLE . ' (`name`,`address`,`description`) 
+        VALUES (:name,:address,:description)');
+
+        $statement->bindValue(':name', $data['name'], \PDO::PARAM_STR);
+        $statement->bindValue(':address', $data['address'], \PDO::PARAM_STR);
+        $statement->bindValue(':description', $data['description'], \PDO::PARAM_STR);
+        $statement->execute();
+    }
+}
