+++ src/Model/RoomManager.php
@@ -35,16 +35,7 @@ class RoomManager extends AbstractManager
         return $this->pdo->query('SELECT *, r.id FROM room r JOIN address ON r.address_id = address.id ORDER BY 
         address.id ASC, r.area ASC')->fetchAll();
     }
-    public function insertAddress(array $data): void
-    {
-        $statement = $this->pdo->prepare('INSERT INTO address (`name`,`address`,`description`) VALUES
-         (:name,:address,:description)');
 
-        $statement->bindValue(':name', $data['name'], \PDO::PARAM_STR);
-        $statement->bindValue(':address', $data['address'], \PDO::PARAM_STR);
-        $statement->bindValue(':description', $data['description'], \PDO::PARAM_STR);
-        $statement->execute();
-    }
     public function selectAddress(): array
     {
         return $this->pdo->query('SELECT * FROM address ORDER BY
